2048 and Pacman with neural nets
2048 will begin if you click new game. Pacman begins if you hit the left
arrow. 2048 works well and I got a high score of around 6000. Pacman
still has some issues... run it on the practice mode and it works fine
as their are infinite lives; however, I need to put a system in place so
the neural net pauses during the brief time after you die and before you
revive so the neural net isn't trained on the null inputs.